Reduce chance for command output overflow
When an error occurs, the caller may leave a result in the input buffer. If
that result is large enough, and the new computation is slow enough, that
result may be used for the next command, which can put us into a vicious
loop where each command reads the output from the previous command!

Avoid this by more aggressively clearing the input buffers with a
non-blocking read before writing a new command.